refactor: price라는 값 객체를 모든 도메인에서 공통으로 사용하도록 변경
